backend/requirements2.txt
Flask==3.0.2
python-dotenv==1.0.1
mysql-connector-python==8.3.0
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
tabulate==0.9.0 